+++ apps/console/src/hooks/graph/PeopleGraph.ts
@@ -63,56 +63,3 @@ export function usePeople(
     return data.organization?.profiles?.edges.map(edge => edge.node) ?? [];
   }, [data]);
 }
-
-export const paginatedPeopleQuery = graphql`
-  query PeopleGraphPaginatedQuery($organizationId: ID!) {
-    organization: node(id: $organizationId) {
-      id
-      ... on Organization {
-        ...PeopleGraphPaginatedFragment
-      }
-    }
-  }
-`;
-
-export const paginatedPeopleFragment = graphql`
-  fragment PeopleGraphPaginatedFragment on Organization
-  @refetchable(queryName: "PeopleListQuery")
-  @argumentDefinitions(
-    first: { type: "Int", defaultValue: 50 }
-    order: {
-      type: "ProfileOrder"
-      defaultValue: { direction: ASC, field: FULL_NAME }
-    }
-    filter: { type: "ProfileFilter", defaultValue: null }
-    after: { type: "CursorKey", defaultValue: null }
-    before: { type: "CursorKey", defaultValue: null }
-    last: { type: "Int", defaultValue: null }
-  ) {
-    canCreatePeople: permission(action: "iam:membership-profile:create")
-    profiles(
-      first: $first
-      after: $after
-      last: $last
-      before: $before
-      orderBy: $order
-      filter: $filter
-    ) @connection(key: "PeopleGraphPaginatedQuery_profiles") {
-      __id
-      edges {
-        node {
-          id
-          fullName
-          emailAddress
-          kind
-          position
-          additionalEmailAddresses
-          contractStartDate
-          contractEndDate
-          canDelete: permission(action: "iam:membership-profile:delete")
-          canUpdate: permission(action: "iam:membership-profile:update")
-        }
-      }
-    }
-  }
-`;
